Add route for edit program
